refactor: use useGetProject instead of project provider to remove coupling with pages
