+++ web_ui/src/core/models/hooks/use-models.hook.ts
@@ -9,11 +9,11 @@ import isEmpty from 'lodash/isEmpty';
 import { useProjectIdentifier } from '../../../hooks/use-project-identifier/use-project-identifier';
 import { NOTIFICATION_TYPE } from '../../../notification/notification-toast/notification-type.enum';
 import { useNotification } from '../../../notification/notification.component';
-import { useTasksWithSupportedAlgorithms } from '../../../pages/project-details/hooks/use-tasks-with-supported-algorithms';
 import { ProjectIdentifier } from '../../projects/core.interface';
 import QUERY_KEYS from '../../requests/query-keys';
 import { useApplicationServices } from '../../services/application-services-provider.component';
 import { getErrorMessage } from '../../services/utils';
+import { useTasksWithSupportedAlgorithms } from '../../supported-algorithms/hooks/use-tasks-with-supported-algorithms';
 import { TrainingBodyDTO } from '../dtos/train-model.interface';
 import { ModelGroupIdentifier, ModelIdentifier, ModelsGroups } from '../models.interface';
 import { ModelDetails } from '../optimized-models.interface';
